feat: implement postinstall script to download 'dist' folder from S3 if not already present
