+++ package.json
@@ -11,7 +11,8 @@
     "lint": "vue-cli-service lint",
     "dev": "nodemon server/index.js",
     "prod": "node server/index.js",
-    "start": "node server/index.js"
+    "postinstall": "node server/utils/postinstall.js",
+    "start": "npm run postinstall && node server/index.js"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.995.0",
